vboxsf: Fix vboxsf_stat in sf_fill_super failing when build with gcc 9
strlcpy to the 1 byte sized shfl_string.string.utf8 array does not work
as expected (with gcc 9) even though there is place for 2 bytes in the
string union, only 1 byte gets set and the string is no 0 terminated.

Simply directly set string.utf8[0] and string.utf8[1] to work around this.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>