vboxsf: Keep a list of open file-handles in the inode
Before this commit sf_reg_open was storing a pointer to the struct
file from the last open in sf_i->file and sf_reg_close would set
sf_i->file to NULL.

This means that if after a RW open of the inode, a RO only open was
done, or a close happens, that sf_i->file would point to a RO handle,
resp. be NULL, either of which will break any (further) writepage calls
done through a shared, RW, mmap of the earlier RW open.

This commit fixes this by keeping a list of open file-handles in the inode
and walking that to get a writeable handle in writepage.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>